# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Incident Ticketing API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CategoryLookup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    PriorityLookup:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sla_response_hrs:
          type: integer
        sla_resolve_hrs:
          type: integer
    StatusLookup:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
paths:
...
  /lookups/categories:
    get:
      summary: List categories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CategoryLookup"
  /lookups/priorities:
    get:
      summary: List priorities
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Priority list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PriorityLookup"
  /lookups/statuses:
    get:
      summary: List ticket statuses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Status list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/StatusLookup"
...
  /tickets/{id}/comments/:
    post:
      summary: Add comment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Comment created
    get:
      summary: List comments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Comment list
  /tickets/{id}/audit:
    get:
      summary: Get ticket audit logs
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Audit logs
...